feat: add event detail page with routing, UI polish, and animations
