+++ src/components/EventCard.jsx
@@ -1,5 +1,5 @@
 import { getEventStatus } from "../utils/eventHelpers";
-
+import { Link } from "react-router-dom";
 export default function EventCard({ event }) {
   const status = getEventStatus(event.date);
   const formattedDate = new Date(event.date + "T00:00:00").toLocaleDateString(
@@ -19,57 +19,51 @@ export default function EventCard({ event }) {
   };
 
   return (
-    <article className="event-card" id={`event-${event.id}`}>
-      <div className="event-card__header">
-        <span className="event-card__category">{event.category}</span>
+    <Link
+      to={`/event/${event.id}`}
+      style={{ textDecoration: "none", color: "inherit" }}
+    >
+      <article className="event-card" id={`event-${event.id}`}>
+        <h2 className="event-card__title">{event.title}</h2>
+        <p className="event-card__description">{event.description}</p>
 
-        {status !== "none" && (
-          <div className={`status-badge ${statusMap[status]}`}>
-            {status === "live" && <span className="live-dot" />}
-            {status === "live" ? "Live Now" : status}
+        <div className="event-card__meta">
+          <div className="event-card__meta-item">
+            <span className="event-card__meta-icon">📅</span>
+            <span>{formattedDate}</span>
+          </div>
+          <div className="event-card__meta-item">
+            <span className="event-card__meta-icon">🕐</span>
+            <span>{event.time}</span>
+          </div>
+          <div className="event-card__meta-item">
+            <span className="event-card__meta-icon">📍</span>
+            <span>{event.location}</span>
+          </div>
+        </div>
+
+        {event.tags && event.tags.length > 0 && (
+          <div className="event-card__tags">
+            {event.tags.map((tag) => (
+              <span key={tag} className="event-card__tag">
+                #{tag}
+              </span>
+            ))}
           </div>
         )}
-      </div>
 
-      <h2 className="event-card__title">{event.title}</h2>
-      <p className="event-card__description">{event.description}</p>
-
-      <div className="event-card__meta">
-        <div className="event-card__meta-item">
-          <span className="event-card__meta-icon">📅</span>
-          <span>{formattedDate}</span>
-        </div>
-        <div className="event-card__meta-item">
-          <span className="event-card__meta-icon">🕐</span>
-          <span>{event.time}</span>
-        </div>
-        <div className="event-card__meta-item">
-          <span className="event-card__meta-icon">📍</span>
-          <span>{event.location}</span>
-        </div>
-      </div>
-
-      {event.tags && event.tags.length > 0 && (
-        <div className="event-card__tags">
-          {event.tags.map((tag) => (
-            <span key={tag} className="event-card__tag">
-              #{tag}
-            </span>
-          ))}
-        </div>
-      )}
-
-      {event.url && (
-        <a
-          href={event.url}
-          className="event-card__link"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn more
-          <span className="event-card__link-arrow">→</span>
-        </a>
-      )}
-    </article>
+        {event.url && (
+          <a
+            href={event.url}
+            className="event-card__link"
+            target="_blank"
+            rel="noopener noreferrer"
+            onClick={(e) => e.stopPropagation()} // 🔥 important fix
+          >
+            Learn more →
+          </a>
+        )}
+      </article>
+    </Link>
   );
 }
