+++ src/index.css
@@ -1422,8 +1422,24 @@ body {
   color: var(--text-primary);
 }
 
+a:focus {
+  outline: 2px solid #8b5cf6;
+  outline-offset: 2px;
+}
+
 @media (max-width: 768px) {
   .sponsors {
     padding: 3rem 1rem;
+  }
+}
+
+@keyframes fadeInUp {
+  from {
+    opacity: 0;
+    transform: translateY(12px);
+  }
+  to {
+    opacity: 1;
+    transform: translateY(0);
   }
 }
